requirements.txt
cycler==0.10.0
imgviz==1.1.0
kiwisolver==1.2.0
labelme==4.4.0
matplotlib==3.2.1
numpy==1.18.5
Pillow==7.1.2
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
PyYAML==5.3.1
QtPy==1.9.0
six==1.15.0
termcolor==1.1.0
tqdm==4.46.1
